Don't space pad parameters with non-zero scale
If a bound parameter has non-zero scale then don't force the
width to the declared precision when converting it to a string
as that will cause values with less digits to have space padding
at the left hand side.

That means that the following code that looks for a sign will
fail to detect it and worse it seems that mysql_bind_param will
ignore negative values if they have leading space.