ODBC-158 The fix and the testcase.
Those aggregate functions, that caused error, returned LONGLONG value,
and Access was getting it as SQL_C_LONG. The problem was that connector
returned lenght of the data as 8(field's size), and not 4(requested
C-type size), as specs prescribe.
Changed paranoid code of length calculation to more sane paranoid code.