ODBC-159 The fix and the testcase
In case if the query is a multistatement, connector splitted it into
individual statements, prepare one by one, and then execute one by one.
In case if one of statements depends on execution of one of previous statements,
application can get error(think of tmp table creation with
select from it) or incorrect results. We can't do anything in case of
SQLPrepare + SQLExecute. The patch fixes that for SQLExecDirect. Now
connector prepares and immediately executes each statement in the batch.