Fix of memory leaks in the driver.
Query structure delete routine was called in the wrong place for
SQL_DROP, and thus wasn't invoked for single statement queries.
Other leak would only occur in case of multistatement query preparing
and batches are not allowed for this connection.
One more leak would occur if sequence of statemence optimized to be
executed via text protocol.
Also the fix of the memory leak in tests. Not really a problem, but created noise in valgrind output.