ODBC-182 The fix and the testcase
If parameter SQL_TIME field was bound as SQL_C_TIMESTAMP, and day field
was not zero, the inserted time value would be different from the value
in time fields of the parameter(server would add total number of hours in those days to the time). The patch makes connector to copy only time fields for the parameter.
Also, the patch enforces time and date validity checks for such parameters, as the specs require.