ODBC-169 The fix and the testcase.
If data was fetched using SQLGetData, with batches of SELECTS that would fail like described int the bug - empty values or even crashes. The reason was that in such case one of structures involved in data fetching was not reset on move to the new resultset. That caused aforementioned errors.
Also the patch fixes SQLRowsCount for batches of upserts or other
statements generation affected rows count.
Removed one unused funxtion, which also relied on C/C internals