ODBC-192 The fix and the testcase + ODBC-194 test
The problem was incorrect buffer address calculation for TIMESTAMP type in case of row-based columns binding. That caused error in ADO, and could cause a crash.

Also commit contains testcase for ODBC-194 with problematic place
commented out.

Also some possible issues are fixed along the way. Like caring of the
case when value buffer is not provided for a field. Or using if
Indicator and StrLen (for the column) have different buffers.