ODBC-194 and ODBC-197 - fixes and testcases
Connector would not return NULL for 0000-00-00 datetime values in case
of SQLGetData call, while doing that in SQLFetch.
Also it would not do that ifn case of empty string conversion to
date/time types.
If time field fetched as timestamp type, fractional part was set to 0.

The patch makes SQLFetch and SQLGetData to use the same function to copy
data to application buffers and process erroneous values.