ODBC-193 Enabling connect and most of operations
SQLWCHAR encoding is depending on DM or on the size of SQLWCHAR. It's either utf16(UnixODBC) or utf32(iODBC). Since iconv needs to know if utf32 le or be, defining charset info struct for le order of bytes. More like a hack to make mariadb_coonvert_string to give iconv correct encoding name.
Fixed failing testcase.