Fix of connector and testcases to make them work
everywhere(Windows, UnixODBC, Travis, Appveyor, iOdbc) - not guearantee
that it really does it, but approaches that goal for sure.
Fix of bug in conversion of utf8 to utf16 if utf8string contains character(s) resquiring 2 SQLWCHAR units in utf16 encoding.
Change of SQL_MAX_COLUMN_NAME_LEN to be 255(256-1), since that is
effectively the longest name len that c/c provides