ODBC-219 The fix and the testcase.
The bug boils down to not reporting correct columns count in case of a query
with multiple results(stored procedure or statements batch) if the result
with affected rows count followed a resultset.
In case of stored procedure(like in the bug report) this is always the case, if
SP returns a result set.
The bug occured becaule IRD was not reset in the SQLMoreResults in the
described case.
Fixed mistake in previous commit(for ODBC-216) - lost statement handler
reinitialization in one place.
Changed tests framework to print test result only after Stmt handler
provided by framework to tests, reinitialized.