ODBC-231 The fix and the testcase.
The bug is rather in SSIS, I would say. It binds column size as signed
int, but the value for LONGTEXT is max unsigned int. Connector was
returning truncation error on the row fetch. And that is probably not
quite right anyway, from everything I read is specs. Anyway, the patch
makes connector not to return truncation error in case if truncation is
caused solely by sign-ness of the field/buffer.