ODBC-251 the testcase only + ODBC-210
It's not quite clear if we should fix anything with 251, but adding the
testcase to have it.
Fixed rest of compilation warnings with iOdbc. At some point I've
decided it's better to suppress the most common warnings in tests - in
iOdbc SQLCHAR is unsigned char, and that caused tons of wranings
Paramethrised some OSX packaging files along the way and fixed some
cmake some found potential cmake issues.