ODBC-267 Cleaned setup library build and linking
Connection test now uses ODBC API, and not driver internals. Driver is
not linked directly. Removed dependencies on the C/C as well DSN info from the dialog is passed to the
connection API in the string, and not in the internal DSN structure.
Added connection string to the connetion test result window.