ODBC-280 The fix of possible crash in threaded environment.
The crash could occur if 2 connections established at same time in
different threads. One of functions read/wrote to the same memory, and
it's call wasn't guarded. The fix actually completely removed it from
connection routine, because it's not really needed to be re-run for each
connecion. Now it part of env init, and also the call is guarded.
Some warnings fixes on Windows.