Build the MariaODBC driver as a soname-less module
When built as a module instead of a shared library, the library is
no longer exporting an unversioned soname. Building the library with
an unversioned soname is superfluous as the linker will use the
filename for linking. The soname is only necessary when the library
linked during build has different filename than the library to be
used at runtime.