ODBC-297 MS Access lists tables from all databases
And errors if table from the database, different from the current, is
selected. That happened because of a bit ambiguous description in the
ODBC specs, what should be done if catalog/schema parameter is NULL. But
we should trust MS's interpretations of own specs. Apparently only
current DB tables are expected in this case.

Update of C/C to v3.1.11

Optimization of some internal structures size.

Changed in travis config server versions to test against