Made possible not to mark certain unstable test(s)
If TEST_SKIP_UNSTABLE_TEST env variable is set one test in connstring is
marked as unstable, and then framework doesn't count it as failed. The
test is known to fail sometimes because of the bug in UnixODBC. Since that testsuite is
to ensure, that dsn is written  an read correctly, that is not critical.
We are fine if it succeeds with one DM at least.