ODBC-330 Added WiX as prerequisite in BUILD.md and check of WiX binaries
If they are not found in given location, and WIX_DIR was provided by
user, the error is returned. If they are not found at location pointed
by env WIX variable, then msi build targets are not generated, but no
error returned.
Fixed error, that WIX_DIR had to have slash or backslash at the end.