ODBC-298 Added NULLISCURRENT connection string option
It controls how is interpreted NULL value in catalog parameter. If it's
not set, then current behavior is used, i.e. that means any schema.
Otherwise it means currently selected database. The patch includes
testcase extension to cover this option.

Addition to ODBC-352 fix - the windows build was broken, if plugins set
to static, and that is the default value.

Slightly improved FindDM.cmake

Moved /usr/lib/${CMAKE_LIBRARY_ARCHITECTURE} from searched paths for
32bit architectures only, to all architectures. It's probably an
overkill anyway(as ${CMAKE_LIBRARY_ARCHITECTURE} is already in the
list of searched suffixes), but more consistent overkill.