ODBC-359 add NULL check to buffer in MADB_FixFetchedValues case SQL_C_WCHAR
It is possible to pass a valid length buffer but no target buffer (and set BufferLength to 0) in SQLBindCol
to get the length of that columns data but not actually copy it. To check if an error occurred the ptr was dereferenced without
first checking if it points to a valid buffer which lead to a segfault.