ODBC-369 Option for result set streaming
aka for not caching the whole RS on the client, but fetch it row by row
from the server. The name of the option is NO_CACHE or STREAMRS is the
alias, 1048576 is the OPTION's bit. The feature only comes in action with
FORWARD_ONLY cursors. Thus it will always be used if combined with FORWARDONLY
option. Some features cannot work in combination with this feature, and
the error HY000 with the message "The requested operation cannot be
executed at the moment...". In particular positioned operation on open
cursor cannot be performed, if streaming is used for cursor.
Tests on travis on linux, but not against skysql, are run twice - in
"normal" mode, and with rs streaming on.
Added to the struct defining test the field for callback deciding if the
test has to be skipped.