ODBC-380 Memory leak during connect if multistatement option selected
The leak occurred once per connection due to missing
MADB_DynstrFree call(the other small change in commit is not related).

Some cleaning in the testcases, so that their own leaks do not create
any noise.

Fix of the memory leak, if string field is SQLGetData'ed as datetime,
In SQLColumns in case of invalid parameters error,
In most cases of SQLGetData as SQL_NUMERIC,
In execution with parameters array for date/time parameter types.