ODBC-397 Introducing LRU srver side prepared statement cache
The cache is used by default. It's use can be turned off by setting
cache size to 0. The size of cache can be set using connection string
option PSCACHESIZE. The default value is 250
Another option is MAXCACHEKEY - it limits the size of key used to cache
SSPS. Basically it's the length of query + lenght of Schema name + 1