ODBC-163 Support of callbacks
In this commit the part(not complete, but most of it - options is
missing, tests changes etc) for parameter callbacks only.
They are used for parameter arrays operations aka bulk commands.
Connector use separate callbacks for parameter columns(kinda extension
to what c/c does/allows) in the c/c++ subsystem of c/odbc. That allows
to have simple, short and easy to read callbacks.