ODBC-424 3.2.2 won't connect to MySQL server on Windows
There was error in the driver, that caused taking substring longer than
total string length. On Windows that caused exception.
Additionaly fixed the way ANSI_QUOTES sql mode is tracked with MySQL
status, as it is not present in the status, and session tracking is
required. But that is a minor thing anyway.