ODBC-426 SQLForeignKeys won't work with MySQL with some parameter
combinations. If foreign key schema parameter is provided and primary
key sckema is not, the call to SQLForeignKeys would result in error.
Also fixsd small bug in SQLProcedureColumns - metadata for double and
float type could be not quite correct, or rather not in line with what
would it be with MariaDB server.