ODBC-427 Out parameters wasn't written with MySQL server
Driver could not detect if the RS is out parameters in case of MySQL
server. The fix caches resultset, fetches next, and if it is the last
one - assumes the current(cached) resultset is out parameters.
Some more fixes of testcases for MySQL

Fixed travis script for testing against MySQL server to give a hint
where to look for authentication plugins