ODBC-438 Driver failed to close PS if cache is used
The cache logic was fine, the problem was in the function that closes
statements. In some (not covered by the old test) cases it could never
obtain the lock, ant thus did not close the statement.
Added new testcase and also chaged the old PS cache test to cover this
case as well.