odbc-443 Wrong result by SQLGetData for unicode string with surrogate
pair. The function used internaly length in characters, and simply
multiplied it by the sizeof(SQLWCHAR) to calculate length in bytes. That
caused results to be read wrongly in particular by ADO as the driver
reported wrong number of available bytes.
C/C has been moved to latest commit that is supposed to be 3.4.2