ODBC-447 Crash on FillSchema command in .Net
Somehow .Net in that case sends SQLMoreResults w/out actually executing
the statement. That comes unexpected for the driver and it crashes. It
looks like good case for DM to return "HY010 Function sequence error",
but it does not, and specs doesn't list it for this error. So,
SQL_NO_DATA_FOUND must be and no crash for sure.