ODBC-448 Converting int fields to double should not cause error
The driver used to return truncation error if the value could not be
converted exactly. It seems like correct specs interpretation would be
not to do so as long as the value in the range of the double(or float)
values. All int values are in that range.
Mover the C/C to v3.3.13