ODBC-405 Error on reading server decimal variable in ADO
Because of MDEV-34179 the reported size of the column in case of
variable is bigger, than the maximum precision for decimal type we
return in the SQLGetTypeInfo. And that freaks ADO out and it returns
error.
The fix limits the precision of the decimal column to max precision of
the decimal column, i.e. to the value the connector returns in the
SQLGetTypeInfo.
The test to ensure that has been added.
(cherry-picking from 3.2)