Fix of the build for not ODBC 3.8 DM's(read iOdbc) and new header
packaging.

In the commit introdusing mariadb specific attributes were also added
checks and correct sql states for unknown attributes from DS specific
range. This used definition that is not present in iOdbc
Since iOdbc does not really support ODBC 3.8 that header is not really
needed. But for the case of direct linking it may be still useful.