ODBC-458 part1, and possibly the last for now. Fixing NOBIGINT option
After initial implementation it lacked one thing - SQLColumns still
returned correct type. That worked for the initial issue, but not for
this one with MS Access. MS Access draw its conclusions on what to do
with bitint columns based on what SQLColumns returs for the data type.

This commit fixes that and extends the testcase to cover this as well.
Had to re-organize completely SQLProcedureColumns, as it got affected as
well. Also, make it to use the same macro as SQLColumns for DATA_TYPE
column, and removed its very own macros.