ODBC-459 MS Access shows record with auto_increment as deleted
right after insert.
To read(and verify) the newly inserted record MSAccess uses the query
with "WHERE `pkField` IS NULL in this case and unlike in old good times -
it does this even 3 times and thus auto_is_null is not of much help
here. Besides it looks like setting SQL_AUTO_IS_NULL did not work - the
patch fixes this as well.
It appears that there is no better solution, than to substitute the
query fitting the pattern after insert id generation. This is done for
Access on Windows only. IS NULL changed to =<ID>.