ODBC-458 Fix of working with bigint auto_increment field in MSAccess
The Access would mark all rows in such field as #Deleted, as it probably
somehow read the values differently in different queries it uses, and
comparison failed. The fix changed default field type to SQL_VARCHAR in
case the driver detects it works with Access. It can do it only on the
Windows atm.