ODBC-462 Client side batch did not respect if multistatement disallowed
Semicolon separated batch could be created if multistatement option was
not selected. That would cause the operation to yield the error.
The patch changes a bit the design since the class did could not get
that information, and passing it there as parameter did not look good.
Besides it now look much more like it's done in c/c++