ODBC-464 Removing possible race conditions in SQLCancel
The way SQLCancel did locking could result in statement structure to be
shared between threads(executing and canceling).
Besides it was not completely following specs. That is improved, but
some behavior is preserved and completion is left for ODBC-465.
So, for ODBC2 driver closes cursor, but killing on the server is
preserved. For ODBC3 it checks if statement's execution needs data and
resets that. Otherwise it still either closes cursor(to preserve current
behavior), but also sets the flag for "delayed" cancelling(if executing
thread gets to executing later than cancel acquires lock). Or it kills
the query at server.