ODBC-466 Making sure that SQLCancel does not run on freed statement
The fix is introducing of the additional global lock and the global list
of deleted statement handles.
Now SQLCancel and SQLFreeStmt(SQL_DROP) first take that global lock.
SQLCancel checks if the statement is not deleted and returns if so. Or
takes over the special lock in the statement.
The list is limited on the nuber of entries with older entries being
removed first. Also, newly allocated Stmt handle checks and removes the
entry holding same id. The list is freed when the last env handle is
freed or when the library is unloaded.