Fix of error in case of RS streaming and fetching data with SQLGetData
The driver could issue extra fetch command if more than one field data
was fetched with SQLGetData. The but was observed in case of binary
protocol(SQLPrepare + SQLExecute in default settings), but I did not
verify if the same would happen with text protocol.

Fixed destructors of some internal classes - the methods they called
could throw.

Turned off testing in RS straming mode in github as more tests require
attention in this mode.