ODBC-471 Fetching of stored procedure out parameters could fail
in case of RS streaming.
The cursor is always forward only in case of resultset streaming,
but out parameters resultset reading requires to move the cursor
back after writing values to the bound parameter buffers. That caused error.
The fix sets flag in such resultsets and allows the operation for
reslultset with the flag.

ODBC-471(partly) binary resultset called function that could throw in
some conditions.