Fix of how parameter codec process insdicator values.
Now it's processed by single coded set for the whole row as it is not
individual for different column/codec types. It checks and sets
inddicator values, e.g. in case of NULL value indicator.
This row-level codec is now set always.
Slightly optimized how descriptor values are read in codecs.