ODBC-473 SQLGetData could write data of the wrong length in case of
resultset streaming and cached resultset. Cached here means - cached in
the ODBC driver, not on the client side in general. That caching occurs
if not the whole result has been fetched, and other statement is
executed.

Also contains more fixes of the testcases for the RS streaming and other
small fixes.

Turned on testing in RS streaming mode in github actions. Added in
github actions pause after driver installation on windows. Apparently
it's not always finished at the moment of DSN creation.