ODBC-469 certificate is always verified it the conncetion test in dialog
Certificate verification in 3.2 driver has 3rd state. This default state
makes it on in case of encrypted connection and off otherwise. But in
the dialog connection test function it was impossible to tell if it's
off or in 3rd state and thus to disable certificate verfification.

The fix introduced for the SSLVERIFY special type of boolean option and
make it use combobox with 3 choices - auto, off and on.