ODBC-474 Correction for the fix
The initial fix was in the commit "Fix of how parameter codec process
insdicator values." 62010b294fcbeadc17e760b590a433feaefb0fc2

It had a bug with moving indicator pointers to the next row and not
restting NULL indicator value in the bind structure. Also this
commit fixes error that column level colback would not move descriptor
pointers if row was skipped or if NULL value indicator was set.