ODBC-481 Crash in SQLDisconnect after SQLCancel if it only closed cursor
One of locks stayed locked and it is not re-entrant.
Added testcase in separate suite as first I expected other conditions
would be needed to be met in order to re-create the issue - I thought it
would need "clean" and only env. But let it stay separate as that suite
may get another use.